figuring out how to import data from librecalc installed pandas-ods-reader
Getting error:
lxml.etree.XMLSyntaxError: Start tag expected, '<' not found, line 1, column 1

Turns out that pandas-pivot-tables needs the file to not have a password.  It also
turned out that librecalc had a bug that did not remove the password when saving the
file without the password option checked.  Upgraded to the newest version of
librecalc and that solved the problem with the password.

A Guide to Pandas Pivot Table
https://builtin.com/data-science/pandas-pivot-tables